+++ src/constants/palette.ts
@@ -1,1 +1,0 @@
-export const sceneColorTypes = ['primary', 'secondary', 'error', 'warning', 'info', 'success'];